[BZ #16141] strptime: fix %z minutes calculation
This is a fix for [BZ #16141] strptime %z offset restriction.

strptime supports the parsing of a timezone offset from UTC time into the
broken-out time field tm_gmtoff. It supports timezone offsets between
UTC-12:00 and UTC+12:00, returning an error (NULL) for values outside
that range.

However, the minutes portion calculation is correct only for minutes
portion values evenly divisible by 3. This is because the minutes value
is converted to decimal time, and rounding results in an incorrect offset
calculation. For example, a +1159 offset string results in a tm_gmtoff
of 43128 (== 11 * 3600 + 3528) seconds, instead of 43140 (== 11 * 3600 +
59 * 60) seconds.

This fix calculates the offset from the hour and minute portions directly,
without the rounding errors introduced by decimal time.

James

2014-12-02  James Perkins  james@loowit.net

	[BZ #16141]
	* time/strptime_l.c (__strptime_internal): Fix %z minutes
	calculation, removing incorrect decimal time rounding, so that
	all minute values result in a valid seconds value