time/tst-strptime2.c: test full input range -9999 to +9999
strptime's %z specifier parses a string consisting of a sign ('+'
or '-'), two hours digits, and optionally two minutes digits, into
a tm.gmtoff field containing the signed number of seconds the time
zone is offset from UTC time.

The time/tst-strptime2.c file passes a short list of strings through
strptime, validating that either the gmtoff value returned matches
an expected value, or that strptime returns an expected NULL for
invalid strings (for example, when the minutes portion of the string
is outside of the range 00 to 59, or the sign is missing before the
hours digits).

In review of strptime fixes, Carlos O'Donell expressed a wish that
the test function iterate through the entire range of all possible
numeric strings (-9999 to +9999) which could be passed to strptime
%z, and validate the correct response.

Specifically, the test will look for a NULL response from strptime
when:

  * sign ('+' or '-') exists before the first digit (invalid
    format).
  * A sign and no digits are found (invalid format).
  * A sign and one digit are found (invalid format).
  * A sign and three digits are found (invalid format).
  * A sign and four digits (-9999 to +9999) are found but the last
    two digits (minutes) are in the range 60 to 99.

The test will look for a success response from strptime with tm.gmtoff
matching the calculated gmtoff value when:

  * A sign and four digits are found (-9999 to +9999), and the last
    two digits (minutes) are in the range 00 to 59.
  * A sign and two digit strings are found (-99 to +99).

The test's iteration over the possible digit values results in 22218
test runs calculated by the test, run, and passed by strptime:

  * In 12198 of the runs, strptime returns successfully with gmtoff
    matching the expected value.
  * In 10020 of the runs, strptime returns failure (NULL) because the
    string was an invalid format.

2015-08-13  James Perkins  <james@loowit.net>

	* time/tst-strptime2.c (tests): Replace short list of test
	strings for strptime %z specifier with code that calculates
	every combination of digits. Add tests for 0, 1, and 3 digit
	invalid strings, which strptime correctly rejects.